Move TTRCache to utils/.
