+++ app/components/entity_mentions/code_links.py
@@ -6,12 +6,12 @@ from typing import NamedTuple
 import discord
 from zig_codeblocks import highlight_zig_code
 
-from .cache import TTRCache
 from app.components.zig_codeblocks import THEME
 from app.setup import gh
 from app.utils import (
     DeleteMessage,
     MessageLinker,
+    TTRCache,
     create_delete_hook,
     create_edit_hook,
     remove_view_after_timeout,
